joined all recipe-related schemas in one file
