+++ frontend/foodieland-app/schemas/recipes.ts
@@ -45,3 +45,13 @@ export const recipeExtendedSchema = z.object({
     ingredients: z.array(ingredientSchema),
     directions: z.array(directionSchema),
 });
+
+export const shortRecipeSchema = z.object({
+    id: z.string(),
+    name: z.string(),
+    description: z.string(),
+    timeToCook: z.number(),
+    category: z.string(),
+    isLiked: z.boolean(),
+    imageData: z.string().nullable()
+});
